Move Contact into core and add core.Assets
Adds a core.Assets interface providing access to all the asset wrapper
types, which flows.SessionAssets now extends with flow definitions and
query resolution. Contact (with Route, ContactEnvelope and friends) moves
into core holding a core.Assets.

Group.CheckQueryBasedMembership takes a contact again now that both live
in the same package. BroadcastTranslations.ForContact replaces the
flows.TranslationsForContact utility, restoring its pre-split shape.
