+++ core/contact.go
@@ -2,11 +2,22 @@ package core
 
 import (
 	"fmt"
+	"strconv"
+	"time"
 
 	"github.com/go-playground/validator/v10"
+	"github.com/nyaruka/gocommon/dates"
+	"github.com/nyaruka/gocommon/i18n"
+	"github.com/nyaruka/gocommon/jsonx"
+	"github.com/nyaruka/gocommon/urns"
 	"github.com/nyaruka/gocommon/uuids"
 	"github.com/nyaruka/goflow/assets"
+	"github.com/nyaruka/goflow/contactql"
+	"github.com/nyaruka/goflow/envs"
+	"github.com/nyaruka/goflow/excellent/types"
 	"github.com/nyaruka/goflow/utils"
+	"github.com/nyaruka/goflow/utils/obfuscate"
+	"github.com/shopspring/decimal"
 )
 
 func init() {
@@ -69,3 +80,657 @@ func (r *ContactReference) String() string {
 }
 
 var _ assets.Reference = (*ContactReference)(nil)
+
+// ContactID is the ID of a contact
+type ContactID int64
+
+// MaxContactURNs is maximum number of URNs a contact can have
+const MaxContactURNs = 50
+
+// schemes of URNs that aren't tied to a specific channel
+var portableURNSchemes = map[string]bool{urns.Phone.Prefix: true, urns.WhatsApp.Prefix: true}
+
+// Contact represents a person who is interacting with the flow
+type Contact struct {
+	uuid       ContactUUID
+	id         ContactID
+	name       string
+	language   i18n.Language
+	status     ContactStatus
+	timezone   *time.Location
+	createdOn  time.Time
+	lastSeenOn *time.Time
+	urns       URNList
+	groups     *GroupList
+	fields     FieldValues
+	tickets    *TicketList
+
+	// transient fields
+	assets Assets
+}
+
+// NewContact creates a new contact with the passed in attributes
+func NewContact(
+	sa Assets,
+	uuid ContactUUID,
+	id ContactID,
+	name string,
+	language i18n.Language,
+	status ContactStatus,
+	timezone *time.Location,
+	createdOn time.Time,
+	lastSeenOn *time.Time,
+	urns []urns.URN,
+	groups []*assets.GroupReference,
+	fields map[string]*Value,
+	tickets []*Ticket,
+	missing assets.MissingCallback) (*Contact, error) {
+
+	urnList, err := NewURNList(sa.Channels(), urns, missing)
+	if err != nil {
+		return nil, err
+	}
+
+	groupList := NewGroupList(sa.Groups(), groups, missing)
+	fieldValues := NewFieldValues(sa.Fields(), fields, missing)
+	ticketsList := NewTicketList(tickets)
+
+	return &Contact{
+		uuid:       uuid,
+		id:         id,
+		name:       name,
+		language:   language,
+		status:     status,
+		timezone:   timezone,
+		createdOn:  createdOn,
+		lastSeenOn: lastSeenOn,
+		urns:       urnList,
+		groups:     groupList,
+		fields:     fieldValues,
+		tickets:    ticketsList,
+		assets:     sa,
+	}, nil
+}
+
+// NewEmptyContact creates a new empy contact with the passed in name, language and location
+func NewEmptyContact(sa Assets, name string, language i18n.Language, timezone *time.Location) *Contact {
+	return &Contact{
+		uuid:       NewContactUUID(),
+		name:       name,
+		language:   language,
+		status:     ContactStatusActive,
+		timezone:   timezone,
+		createdOn:  dates.Now(),
+		lastSeenOn: nil,
+		urns:       URNList{},
+		groups:     NewGroupList(sa.Groups(), nil, assets.IgnoreMissing),
+		fields:     make(FieldValues),
+		tickets:    NewTicketList(nil),
+		assets:     sa,
+	}
+}
+
+// Clone clones this contact when we need to snapshot as part of parent run summary.
+func (c *Contact) Clone() *Contact {
+	return &Contact{
+		uuid:       c.uuid,
+		id:         c.id,
+		name:       c.name,
+		language:   c.language,
+		status:     c.status,
+		timezone:   c.timezone,
+		createdOn:  c.createdOn,
+		lastSeenOn: c.lastSeenOn,
+		urns:       c.urns.Clone(),
+		groups:     c.groups.Clone(),
+		fields:     c.fields.Clone(),
+		tickets:    NewTicketList(nil), // tickets not included
+		assets:     c.assets,
+	}
+}
+
+// UUID returns the UUID of this contact
+func (c *Contact) UUID() ContactUUID { return c.uuid }
+
+// ID returns the numeric ID of this contact
+func (c *Contact) ID() ContactID { return c.id }
+
+// SetLanguage sets the language for this contact
+func (c *Contact) SetLanguage(lang i18n.Language) { c.language = lang }
+
+// Language gets the language for this contact
+func (c *Contact) Language() i18n.Language { return c.language }
+
+// Country gets the country for this contact..
+//
+// TODO: currently this is derived from their preferred channel or any tel URNs but probably should become an explicit
+// field at some point
+func (c *Contact) Country() i18n.Country {
+	ch := c.PreferredChannel()
+	if ch != nil && ch.Country() != i18n.NilCountry {
+		return ch.Country()
+	}
+
+	for _, u := range c.urns {
+		if u.Scheme == urns.Phone.Prefix {
+			c := i18n.DeriveCountryFromTel(u.Path)
+			if c != i18n.NilCountry {
+				return c
+			}
+		}
+	}
+
+	return i18n.NilCountry
+}
+
+// Locale gets the locale for this contact, using the environment country if contact doesn't have one
+func (c *Contact) Locale(env envs.Environment) i18n.Locale {
+	country := c.Country()
+	if country == i18n.NilCountry {
+		country = env.DefaultCountry()
+	}
+	return i18n.NewLocale(c.language, country)
+}
+
+// Status returns the contact status
+func (c *Contact) Status() ContactStatus { return c.status }
+
+// SetStatus sets the status of this contact (blocked, stopped or active)
+func (c *Contact) SetStatus(status ContactStatus) { c.status = status }
+
+// SetTimezone sets the timezone of this contact
+func (c *Contact) SetTimezone(tz *time.Location) { c.timezone = tz }
+
+// Timezone returns the timezone of this contact
+func (c *Contact) Timezone() *time.Location { return c.timezone }
+
+// CreatedOn returns the created on time of this contact
+func (c *Contact) CreatedOn() time.Time { return c.createdOn }
+
+// LastSeenOn returns the last seen on time of this contact
+func (c *Contact) LastSeenOn() *time.Time { return c.lastSeenOn }
+
+// SetLastSeenOn sets the last seen on time of this contact
+func (c *Contact) SetLastSeenOn(t time.Time) { c.lastSeenOn = &t }
+
+// SetName sets the name of this contact
+func (c *Contact) SetName(name string) { c.name = name }
+
+// Name returns the name of this contact
+func (c *Contact) Name() string { return c.name }
+
+// URNs returns the URNs of this contact
+func (c *Contact) URNs() URNList { return c.urns }
+
+// AddRoute adds a new URN to this contact with optional channel affinity. Returns whether the contact was changed.
+// If the URN is already on the contact this is a no-op - to update channel affinity use SetAffinity or SetRoutes.
+func (c *Contact) AddRoute(urn urns.URN, channel *Channel) bool {
+	if c.HasURN(urn) {
+		return false
+	}
+
+	scheme, path, _, _ := urn.ToParts()
+
+	c.urns = append(c.urns, NewURN(scheme, path, "", channel))
+	return true
+}
+
+// RemoveURN removes an existing URN from this contact
+func (c *Contact) RemoveURN(urn urns.URN) bool {
+	if !c.HasURN(urn) {
+		return false
+	}
+
+	newURNs := make([]*URN, 0, len(c.urns)-1)
+	for _, u := range c.urns {
+		if u.Identity() != urn.Identity() {
+			newURNs = append(newURNs, u)
+		}
+	}
+
+	c.urns = URNList(newURNs)
+	return true
+}
+
+// SetRoutes replaces this contact's URNs with the given routes, each carrying optional channel affinity.
+func (c *Contact) SetRoutes(routes []Route) bool {
+	isSame := func() bool {
+		if len(c.urns) != len(routes) {
+			return false
+		}
+		for i, u := range c.urns {
+			if u.Identity() != routes[i].URN.Identity() || u.Channel != routes[i].Channel {
+				return false
+			}
+		}
+		return true
+	}
+	if isSame() {
+		return false
+	}
+
+	c.urns = URNList{}
+	for _, r := range routes {
+		scheme, path, _, _ := r.URN.ToParts()
+
+		c.urns = append(c.urns, NewURN(scheme, path, "", r.Channel))
+	}
+
+	return true
+}
+
+func (c *Contact) HasURN(urn urns.URN) bool {
+	for _, u := range c.urns {
+		if u.Identity() == urn.Identity() {
+			return true
+		}
+	}
+	return false
+}
+
+// SetAffinity sets the preferred URN/channel for this contact by moving it to the front of the list and returns whether
+// any change was made. If the URN is not already associated with the contact then this is a no-op and returns false.
+func (c *Contact) SetAffinity(urn urns.URN, ch *Channel) bool {
+	oldURNs := c.urns.Clone()
+
+	// find the URN in our list
+	var urnIndex = -1
+	for i, u := range c.urns {
+		if u.Identity() == urn.Identity() {
+			urnIndex = i
+			break
+		}
+	}
+
+	// URN not found
+	if urnIndex == -1 {
+		return false
+	}
+
+	targetURN := c.urns[urnIndex]
+
+	// set the channel - we assume channel and URN are compatible since we received something on them
+	targetURN.Channel = ch
+
+	// move URN to front if not already there
+	if urnIndex > 0 {
+		c.urns = append([]*URN{targetURN}, append(c.urns[:urnIndex], c.urns[urnIndex+1:]...)...)
+	}
+
+	return !oldURNs.Equal(c.urns)
+}
+
+// Fields returns this contact's field values
+func (c *Contact) Fields() FieldValues { return c.fields }
+
+// Groups returns the groups that this contact belongs to
+func (c *Contact) Groups() *GroupList { return c.groups }
+
+// Tickets returns the tickets for this contact
+func (c *Contact) Tickets() *TicketList { return c.tickets }
+
+// Reference returns a reference to this contact
+func (c *Contact) Reference() *ContactReference {
+	if c == nil {
+		return nil
+	}
+	return NewContactReference(c.uuid, c.name)
+}
+
+// Format returns a friendly string version of this contact depending on what fields are set
+func (c *Contact) Format(env envs.Environment) string {
+	// if contact has a name set, use that
+	if c.name != "" {
+		return c.name
+	}
+
+	// otherwise use either id or the highest priority URN depending on the env
+	if env.RedactionPolicy() == envs.RedactionPolicyURNs {
+		return strconv.Itoa(int(c.id))
+	}
+	if len(c.urns) > 0 {
+		return c.urns[0].Encode().Format()
+	}
+
+	return ""
+}
+
+// Context returns the properties available in expressions
+//
+//	__default__:text -> the name or URN
+//	uuid:text -> the UUID of the contact
+//	ref:text -> the reference identifier of the contact
+//	first_name:text -> the first name of the contact
+//	name:text -> the name of the contact
+//	language:text -> the language of the contact as 3-letter ISO code
+//	status:text -> the status of the contact
+//	created_on:datetime -> the creation date of the contact
+//	last_seen_on:any -> the last seen date of the contact
+//	urns:[]text -> the URNs belonging to the contact
+//	urn:text -> the preferred URN of the contact
+//	groups:[]group -> the groups the contact belongs to
+//	fields:fields -> the custom field values of the contact
+//	channel:channel -> the preferred channel of the contact
+//
+// @context contact
+func (c *Contact) Context(env envs.Environment) map[string]types.XValue {
+	var firstName, urn, timezone, lastSeenOn types.XValue
+
+	ref, _ := obfuscate.EncodeID(int64(c.id), env.ObfuscationKey())
+
+	if c.timezone != nil {
+		timezone = types.NewXText(c.timezone.String())
+	}
+
+	preferredURN := c.PreferredURN()
+	if preferredURN != nil {
+		urn = preferredURN.ToXValue(env)
+	}
+
+	names := utils.TokenizeString(c.name)
+	if len(names) >= 1 {
+		firstName = types.NewXText(names[0])
+	}
+
+	if c.lastSeenOn != nil {
+		lastSeenOn = types.NewXDateTime(*c.lastSeenOn)
+	}
+
+	id := types.NewXText(strconv.Itoa(int(c.id)))
+	id.SetDeprecated("contact.id: use contact.ref instead")
+
+	tickets := c.tickets.Open().ToXValue(env)
+	tickets.SetDeprecated("contact.tickets: use ticket instead")
+
+	return map[string]types.XValue{
+		"__default__":  types.NewXText(c.Format(env)),
+		"uuid":         types.NewXText(string(c.uuid)),
+		"ref":          types.NewXText(ref),
+		"name":         types.NewXText(c.name),
+		"first_name":   firstName,
+		"language":     types.NewXText(string(c.language)),
+		"timezone":     timezone,
+		"status":       types.NewXText(string(c.status)),
+		"created_on":   types.NewXDateTime(c.createdOn),
+		"last_seen_on": lastSeenOn,
+		"urns":         c.urns.ToXValue(env),
+		"urn":          urn,
+		"groups":       c.groups.ToXValue(env),
+		"fields":       Context(env, c.Fields()),
+		"channel":      Context(env, c.PreferredChannel()),
+		"id":           id,
+		"tickets":      tickets,
+	}
+}
+
+// Route is a URN paired with the channel that should handle it
+type Route struct {
+	URN     urns.URN
+	Channel *Channel
+}
+
+// ResolveRoutes resolves possible URN/channel routes for sending
+func (c *Contact) ResolveRoutes(all bool) []Route {
+	routes := []Route{}
+
+	for _, u := range c.urns {
+		channel := c.assets.Channels().GetForURN(u, assets.ChannelRoleSend)
+		if channel != nil {
+			routes = append(routes, Route{URN: u.Identity(), Channel: channel})
+			if !all {
+				break
+			}
+		}
+	}
+	return routes
+}
+
+// PreferredURN gets the preferred URN for this contact, i.e. the URN we would use for sending
+func (c *Contact) PreferredURN() *URN {
+	for _, u := range c.urns {
+		if c.assets.Channels().GetForURN(u, assets.ChannelRoleSend) != nil {
+			return u
+		}
+	}
+	return nil
+}
+
+// PreferredChannel gets the preferred channel for this contact, i.e. the channel we would use for sending
+func (c *Contact) PreferredChannel() *Channel {
+	routes := c.ResolveRoutes(false)
+	if len(routes) > 0 {
+		return routes[0].Channel
+	}
+	return nil
+}
+
+// UpdatePreferredChannel updates the preferred channel and returns whether any change was made
+func (c *Contact) UpdatePreferredChannel(channel *Channel) bool {
+	oldURNs := c.urns.Clone()
+
+	// setting preferred channel to nil means clearing affinity on all URNs
+	if channel == nil {
+		for _, u := range c.urns {
+			u.Channel = nil
+		}
+	} else {
+		if !channel.HasRole(assets.ChannelRoleSend) {
+			return false
+		}
+
+		priorityURNs := make([]*URN, 0)
+		otherURNs := make([]*URN, 0)
+
+		for _, urn := range c.urns {
+			// portable URNs can be re-assigned when supported by channel; a business-scoped WhatsApp URN
+			// (BSUID) is scoped to a single business/channel pair though, so it isn't portable
+			if portableURNSchemes[urn.Scheme] && !urns.IsWhatsAppBSUID(urn.Identity()) && channel.SupportsScheme(urn.Scheme) {
+				urn.Channel = channel
+			}
+
+			// If URN doesn't have a channel and is a scheme supported by the channel, then we can set its
+			// channel. This may result in unsendable URN/channel pairing but can't do much about that.
+			if urn.Channel == nil && channel.SupportsScheme(urn.Scheme) {
+				urn.Channel = channel
+			}
+
+			// move any URNs with this channel to the front of the list
+			if urn.Channel == channel {
+				priorityURNs = append(priorityURNs, urn)
+			} else {
+				otherURNs = append(otherURNs, urn)
+			}
+		}
+
+		c.urns = append(priorityURNs, otherURNs...)
+	}
+
+	return !oldURNs.Equal(c.urns)
+}
+
+// ReevaluateQueryBasedGroups reevaluates membership of all query based groups for this contact
+func (c *Contact) ReevaluateQueryBasedGroups(env envs.Environment) ([]*Group, []*Group) {
+	added := make([]*Group, 0)
+	removed := make([]*Group, 0)
+
+	for _, group := range c.assets.Groups().All() {
+		if !group.UsesQuery() {
+			continue
+		}
+
+		qualifies := group.CheckQueryBasedMembership(env, c)
+
+		if qualifies {
+			if c.groups.Add(group) {
+				added = append(added, group)
+			}
+		} else {
+			if c.groups.Remove(group) {
+				removed = append(removed, group)
+			}
+		}
+	}
+
+	return added, removed
+}
+
+// QueryProperty resolves a contact query search key for this contact
+//
+// Note that this method excludes id, ref, group and flow search attributes as those are disallowed
+// query based groups.
+func (c *Contact) QueryProperty(env envs.Environment, key string, propType contactql.PropertyType) []any {
+	if propType == contactql.PropertyTypeAttribute {
+		switch key {
+		case contactql.AttributeUUID:
+			return []any{string(c.uuid)}
+		case contactql.AttributeName:
+			if c.name != "" {
+				return []any{c.name}
+			}
+			return nil
+		case contactql.AttributeLanguage:
+			if c.language != i18n.NilLanguage {
+				return []any{string(c.language)}
+			}
+			return nil
+		case contactql.AttributeURN:
+			vals := make([]any, len(c.URNs()))
+			for i, urn := range c.URNs() {
+				vals[i] = urn.Path
+			}
+			return vals
+		case contactql.AttributeTickets:
+			return []any{decimal.NewFromInt(int64(c.tickets.Open().Count()))}
+		case contactql.AttributeCreatedOn:
+			return []any{c.createdOn}
+		case contactql.AttributeLastSeenOn:
+			if c.lastSeenOn != nil {
+				return []any{*c.lastSeenOn}
+			}
+			return nil
+		default:
+			return nil
+		}
+	} else if propType == contactql.PropertyTypeURN {
+		urnsWithScheme := c.urns.WithScheme(key)
+		vals := make([]any, len(urnsWithScheme))
+		for i := range urnsWithScheme {
+			vals[i] = urnsWithScheme[i].Path
+		}
+		return vals
+	}
+
+	// try as a contact field
+	nativeValue := c.fields[key].QueryValue()
+	if nativeValue == nil {
+		return nil
+	}
+
+	return []any{nativeValue}
+}
+
+var _ contactql.Queryable = (*Contact)(nil)
+
+//------------------------------------------------------------------------------------------
+// JSON Encoding / Decoding
+//------------------------------------------------------------------------------------------
+
+type ContactEnvelope struct {
+	UUID       ContactUUID              `json:"uuid"                validate:"required,uuid"`
+	ID         ContactID                `json:"id,omitempty"`
+	Name       string                   `json:"name,omitempty"`
+	Language   i18n.Language            `json:"language,omitempty"`
+	Status     ContactStatus            `json:"status,omitempty"    validate:"required,contact_status"`
+	Timezone   string                   `json:"timezone,omitempty"`
+	CreatedOn  time.Time                `json:"created_on"          validate:"required"`
+	LastSeenOn *time.Time               `json:"last_seen_on,omitempty"`
+	URNs       []urns.URN               `json:"urns,omitempty"      validate:"dive,urn"`
+	Groups     []*assets.GroupReference `json:"groups,omitempty"    validate:"dive"`
+	Fields     map[string]*Value        `json:"fields,omitempty"`
+	Tickets    []*TicketEnvelope        `json:"tickets,omitempty"`
+}
+
+func (e *ContactEnvelope) Unmarshal(sa Assets, missing assets.MissingCallback) (*Contact, error) {
+	c := &Contact{
+		uuid:       e.UUID,
+		id:         e.ID,
+		name:       e.Name,
+		language:   e.Language,
+		status:     e.Status,
+		createdOn:  e.CreatedOn,
+		lastSeenOn: e.LastSeenOn,
+		assets:     sa,
+	}
+
+	if e.Timezone != "" {
+		var err error
+		if c.timezone, err = time.LoadLocation(e.Timezone); err != nil {
+			return nil, err
+		}
+	}
+
+	if e.URNs == nil {
+		c.urns = make(URNList, 0)
+	} else {
+		var err error
+		if c.urns, err = NewURNList(sa.Channels(), e.URNs, missing); err != nil {
+			return nil, fmt.Errorf("error reading urns: %w", err)
+		}
+	}
+
+	c.groups = NewGroupList(sa.Groups(), e.Groups, missing)
+	c.fields = NewFieldValues(sa.Fields(), e.Fields, missing)
+
+	tickets := make([]*Ticket, len(e.Tickets))
+	for i, t := range e.Tickets {
+		tickets[i] = t.Unmarshal(sa.Topics(), sa.Users(), missing)
+	}
+	c.tickets = NewTicketList(tickets)
+
+	return c, nil
+}
+
+func (c *Contact) Marshal() *ContactEnvelope {
+	e := &ContactEnvelope{
+		Name:       c.name,
+		UUID:       c.uuid,
+		ID:         c.id,
+		Status:     c.status,
+		Language:   c.language,
+		CreatedOn:  c.createdOn,
+		LastSeenOn: c.lastSeenOn,
+		URNs:       c.urns.Encode(),
+		Groups:     c.groups.References(),
+	}
+
+	if c.timezone != nil {
+		e.Timezone = c.timezone.String()
+	}
+
+	e.Fields = make(map[string]*Value)
+	for _, v := range c.fields {
+		if v != nil {
+			e.Fields[v.Field().Key()] = v.Value
+		}
+	}
+
+	e.Tickets = c.Tickets().Marshal()
+
+	return e
+}
+
+// ReadContact decodes a contact from the passed in JSON
+func ReadContact(sa Assets, data []byte, missing assets.MissingCallback) (*Contact, error) {
+	e := &ContactEnvelope{}
+
+	if err := utils.UnmarshalAndValidate(data, e); err != nil {
+		return nil, fmt.Errorf("unable to read contact: %w", err)
+	}
+
+	return e.Unmarshal(sa, missing)
+}
+
+// MarshalJSON marshals this contact into JSON
+func (c *Contact) MarshalJSON() ([]byte, error) {
+	return jsonx.Marshal(c.Marshal())
+}
