+++ core/msg.go
@@ -2,6 +2,7 @@ package core
 
 import (
 	"fmt"
+	"slices"
 	"strings"
 
 	"github.com/nyaruka/gocommon/i18n"
@@ -9,6 +10,7 @@ import (
 	"github.com/nyaruka/gocommon/urns"
 	"github.com/nyaruka/gocommon/uuids"
 	"github.com/nyaruka/goflow/assets"
+	"github.com/nyaruka/goflow/envs"
 	"github.com/nyaruka/goflow/utils"
 )
 
@@ -233,3 +235,42 @@ type BroadcastUUID uuids.UUID
 func NewBroadcastUUID() BroadcastUUID { return BroadcastUUID(uuids.NewV7()) }
 
 type BroadcastTranslations map[i18n.Language]*MsgContent
+
+// ForContact is a utility to help callers select the translation for a contact
+func (b BroadcastTranslations) ForContact(e envs.Environment, c *Contact, baseLanguage i18n.Language) (*MsgContent, i18n.Locale) {
+	// get the set of languages to merge translations from
+	languages := make([]i18n.Language, 0, 3)
+
+	// highest priority is the contact language if it is valid
+	if c.Language() != i18n.NilLanguage && slices.Contains(e.AllowedLanguages(), c.Language()) {
+		languages = append(languages, c.Language())
+	}
+
+	// then the default workspace language, then the base language
+	languages = append(languages, e.DefaultLanguage(), baseLanguage)
+
+	content := &MsgContent{}
+	language := i18n.NilLanguage
+	country := e.DefaultCountry()
+	if c.Country() != i18n.NilCountry {
+		country = c.Country()
+	}
+
+	for _, lang := range languages {
+		trans := b[lang]
+		if trans != nil {
+			if content.Text == "" && trans.Text != "" {
+				content.Text = trans.Text
+				language = lang
+			}
+			if len(content.Attachments) == 0 && len(trans.Attachments) > 0 {
+				content.Attachments = trans.Attachments
+			}
+			if len(content.QuickReplies) == 0 && len(trans.QuickReplies) > 0 {
+				content.QuickReplies = trans.QuickReplies
+			}
+		}
+	}
+
+	return content, i18n.NewLocale(language, country)
+}
