+++ core/testdata/smart_groups.assets.json
