+++ core/testdata/smart_groups.json
