+++ flows/modifiers/affinity.go
@@ -38,7 +38,7 @@ func NewAffinity(urn urns.URN, channel *core.Channel) *Affinity {
 }
 
 // Apply applies this modification to the given contact
-func (m *Affinity) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *flows.Contact, log events.EventLogger) (bool, error) {
+func (m *Affinity) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *core.Contact, log events.EventLogger) (bool, error) {
 	if contact.SetAffinity(m.urn, m.channel) {
 		// if URNs change in anyway, generate a URNs changed event
 		log(events.NewContactURNsChanged(contact.URNs().Encode()))
