+++ flows/modifiers/base.go
@@ -40,7 +40,7 @@ func newBaseModifier(typeName string) baseModifier {
 func (m *baseModifier) Type() string { return m.typ }
 
 // Apply applies the given modifier to the given contact and re-evaluates query based groups if necessary
-func Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, c *flows.Contact, mod flows.Modifier, logEvent events.EventLogger) (bool, error) {
+func Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, c *core.Contact, mod flows.Modifier, logEvent events.EventLogger) (bool, error) {
 	modified, err := mod.Apply(ctx, eng, env, sa, c, logEvent)
 	if err != nil {
 		return false, err
@@ -52,7 +52,7 @@ func Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows
 }
 
 // ReevaluateGroups is a helper to re-evaluate groups and log any changes to membership
-func ReevaluateGroups(env envs.Environment, contact *flows.Contact, log events.EventLogger) {
+func ReevaluateGroups(env envs.Environment, contact *core.Contact, log events.EventLogger) {
 	added, removed := contact.ReevaluateQueryBasedGroups(env)
 
 	// make sure from all static groups are removed for non-active contacts
