+++ flows/modifiers/seen.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/nyaruka/gocommon/jsonx"
 	"github.com/nyaruka/goflow/assets"
+	"github.com/nyaruka/goflow/core"
 	"github.com/nyaruka/goflow/core/events"
 	"github.com/nyaruka/goflow/envs"
 	"github.com/nyaruka/goflow/flows"
@@ -35,7 +36,7 @@ func NewSeen(seenOn time.Time) *Seen {
 }
 
 // Apply applies this modification to the given contact
-func (m *Seen) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *flows.Contact, log events.EventLogger) (bool, error) {
+func (m *Seen) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *core.Contact, log events.EventLogger) (bool, error) {
 	if contact.LastSeenOn() == nil || m.seenOn.After(*contact.LastSeenOn()) {
 		contact.SetLastSeenOn(m.seenOn)
 		log(events.NewContactLastSeenChanged(m.seenOn))
