+++ flows/modifiers/ticket_assignee.go
@@ -37,7 +37,7 @@ func NewTicketAssignee(ticketUUID core.TicketUUID, assignee *core.User) *TicketA
 }
 
 // Apply applies this modification to the given contact
-func (m *TicketAssignee) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *flows.Contact, log events.EventLogger) (bool, error) {
+func (m *TicketAssignee) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *core.Contact, log events.EventLogger) (bool, error) {
 	ticket := contact.Tickets().Find(m.ticketUUID)
 
 	if ticket != nil && ticket.Assignee() != m.assignee {
