+++ flows/modifiers/ticket_close.go
@@ -35,7 +35,7 @@ func NewTicketClose(ticketUUID core.TicketUUID) *TicketClose {
 }
 
 // Apply applies this modification to the given contact
-func (m *TicketClose) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *flows.Contact, log events.EventLogger) (bool, error) {
+func (m *TicketClose) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *core.Contact, log events.EventLogger) (bool, error) {
 	ticket := contact.Tickets().Find(m.ticketUUID)
 
 	if ticket != nil && ticket.Status() != core.TicketStatusClosed {
