+++ flows/modifiers/ticket_reopen.go
@@ -35,7 +35,7 @@ func NewTicketReopen(ticketUUID core.TicketUUID) *TicketReopen {
 }
 
 // Apply applies this modification to the given contact
-func (m *TicketReopen) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *flows.Contact, log events.EventLogger) (bool, error) {
+func (m *TicketReopen) Apply(ctx context.Context, eng flows.Engine, env envs.Environment, sa flows.SessionAssets, contact *core.Contact, log events.EventLogger) (bool, error) {
 	// if there's already an open ticket, nothing to do
 	if contact.Tickets().Open().Count() > 0 {
 		return false, nil
